+++ src/TextShow/Debug/Trace/Generic.hs
@@ -19,28 +19,30 @@ module TextShow.Debug.Trace.Generic (
     , genericTraceTextShowM
     ) where
 
+import qualified Data.Text as TS (Text)
+
 import Generics.Deriving.Base (Generic, Rep)
 
 import Prelude ()
 import Prelude.Compat
 
 import TextShow.Debug.Trace
-import TextShow.Generic (GTextShowT, Zero, genericShowt)
+import TextShow.Generic (GTextShow, genericShowt)
 
 -- | A 'Generic' implementation of 'traceTextShow'.
 --
 -- /Since: 2/
-genericTraceTextShow :: (Generic a, GTextShowT Zero (Rep a)) => a -> b -> b
+genericTraceTextShow :: (Generic a, GTextShow TS.Text (Rep a ())) => a -> b -> b
 genericTraceTextShow = tracet . genericShowt
 
 -- | A 'Generic' implementation of 'traceTextShowId'.
 --
 -- /Since: 2/
-genericTraceTextShowId :: (Generic a, GTextShowT Zero (Rep a)) => a -> a
+genericTraceTextShowId :: (Generic a, GTextShow TS.Text (Rep a ())) => a -> a
 genericTraceTextShowId a = tracet (genericShowt a) a
 
 -- | A 'Generic' implementation of 'traceShowM'.
 --
 -- /Since: 2/
-genericTraceTextShowM :: (Generic a, GTextShowT Zero (Rep a), Applicative f) => a -> f ()
+genericTraceTextShowM :: (Generic a, GTextShow TS.Text (Rep a ()), Applicative f) => a -> f ()
 genericTraceTextShowM = tracetM . genericShowt